Neaten subsetter code, add custom table
