+++ CharacterMap/CharacterMap/Helpers/Subsetter.cs
@@ -1,12 +1,14 @@
+using Microsoft.Graphics.Canvas;
+using Microsoft.Graphics.Canvas.Geometry;
+using Microsoft.Graphics.Canvas.Text;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
-using Microsoft.Graphics.Canvas;
-using Microsoft.Graphics.Canvas.Geometry;
 using System.Numerics;
-using Microsoft.Graphics.Canvas.Text;
+using System.Runtime.ConstrainedExecution;
+using System.Text;
+using System.Xml.Linq;
 
 namespace CharacterMap.Helpers;
 
@@ -82,67 +84,17 @@ public partial class FontGlyph(
 /// </summary>
 public class FontSubsetter
 {
-    public const string SUBSETTER_TAG = "CMSV";
-    public const string SUBSETTER_SIGNATURE = "CMUWP.Subsetter.v1";
-
-    public static bool IsCreatedBySubsetter(CMFontFace font)
-    {
-        if (font?.Face is null)
-            return false;
-
-        byte[] data = font.Face.GetFontTable(SUBSETTER_TAG);
-        return data is not null && Encoding.ASCII.GetString(data).StartsWith(SUBSETTER_SIGNATURE);
-    }
 
     // ---------------------------
     // Big-endian helpers
     // ---------------------------
 
-    static ushort ReadUInt16BE(BinaryReader br)
-    {
-        byte[] b = br.ReadBytes(2);
-        return (ushort)((b[0] << 8) | b[1]);
-    }
-
-    static short ReadInt16BE(BinaryReader br)
-    {
-        byte[] b = br.ReadBytes(2);
-        return (short)((b[0] << 8) | b[1]);
-    }
-
-    static uint ReadUInt32BE(BinaryReader br)
-    {
-        byte[] b = br.ReadBytes(4);
-        return ((uint)b[0] << 24) | ((uint)b[1] << 16) | ((uint)b[2] << 8) | b[3];
-    }
-
-    static void WriteUInt16BE(BinaryWriter bw, ushort v)
-    {
-        bw.Write((byte)(v >> 8));
-        bw.Write((byte)(v & 0xFF));
-    }
-
-    static void WriteUInt32BE(BinaryWriter bw, uint v)
-    {
-        bw.Write((byte)(v >> 24));
-        bw.Write((byte)((v >> 16) & 0xFF));
-        bw.Write((byte)((v >> 8) & 0xFF));
-        bw.Write((byte)(v & 0xFF));
-    }
-
-    static void WriteInt16BE(BinaryWriter bw, short v)
-    {
-        bw.Write((byte)((v >> 8) & 0xFF));
-        bw.Write((byte)(v & 0xFF));
-    }
-
-    //private struct TableRecord
-    //{
-    //    public string Tag;
-    //    public uint Checksum;
-    //    public uint Offset;
-    //    public uint Length;
-    //}
+    static ushort ReadUInt16BE(BinaryReader br) => SfntWriter.ReadUInt16BE(br);
+    static short ReadInt16BE(BinaryReader br) => SfntWriter.ReadInt16BE(br);
+    static uint ReadUInt32BE(BinaryReader br) => SfntWriter.ReadUInt32BE(br);
+    static void WriteUInt16BE(BinaryWriter bw, ushort v) => SfntWriter.WriteUInt16BE(bw, v);
+    static void WriteUInt32BE(BinaryWriter bw, uint v) => SfntWriter.WriteUInt32BE(bw, v);
+    static void WriteInt16BE(BinaryWriter bw, short v) => SfntWriter.WriteInt16BE(bw, v);
 
     private class FontTableCache : IDisposable
     {
@@ -157,15 +109,8 @@ public class FontSubsetter
         }
     }
 
-    // ---------------------------
-    // Main subsetting function
-    // ---------------------------
-
-
-
     public static async Task<StorageFile> CreateSubsetAsync(SubsetOptions opts)
     {
-        string desiredFamilyName = opts.DesiredName;
         IList<FontGlyph> characters = opts.Characters;
 
         if (characters == null || characters.Count == 0)
@@ -241,7 +186,7 @@ public class FontSubsetter
             outputCmap[unicode] = (uint)(i + 1);
         }
 
-        byte[] newCmapTable = BuildCmapTable(outputCmap);
+        byte[] newCmapTable = SfntWriter.BuildCmapTable(outputCmap);
 
         // 6. Start writing tables
         Dictionary<CMFontFace, FontTableCache> tableCache = [];
@@ -357,265 +302,25 @@ public class FontSubsetter
                 }
             }
 
-            // 6.2. Rebuild Name table with new copyright and descriptions
-            List<CMFontFace> metadataFonts = uniqueFonts
-                .GroupBy(f => $"{f.FamilyName} {f.PreferredName}".Trim())
-                .Select(g => g.First())
-                .ToList();
+            CMSVTable cmTable = CMSVTable.TryDecode(templateFont) ?? new();
 
-            string MergeField(CanvasFontInformation info)
-            {
-                List<string> distinctValues = metadataFonts
-                    .Select(f => f.TryGetInfo(info)?.Value ?? string.Empty)
-                    .Distinct()
-                    .ToList();
+            // 6.2. Merge OS/2 embedding rights and ranges
+            OS2Metadata os2Meta = SfntWriter.CalculateMergedOS2(uniqueFonts, fontFsTypes, templateFont);
 
-                if (distinctValues.Count == 1)
-                    return distinctValues[0];
+            // 6.3. Rebuild tables to match output
+            byte[] newNameTable = SfntWriter.RebuildNameTable(opts, templateFont, uniqueFonts, cmTable);
+            byte[] headData = SfntWriter.RebuildHeadTable(opts, templateFont);
+            byte[] os2Data = SfntWriter.BuildOS2Table(templateFont, os2Meta);
+            byte[] postData = SfntWriter.BuildPostTable(characters, outputNumGlyphs);
+            byte[] newLocaData = SfntWriter.BuildLocaTable(outputNumGlyphs, newLoca);
 
-                List<string> lines = new();
-                foreach (CMFontFace font in metadataFonts)
-                {
-                    string val = font.TryGetInfo(info)?.Value;
-                    string fontName = $"{font.FamilyName} {font.PreferredName}".Trim();
-                    if (!string.IsNullOrEmpty(val))
-                        lines.Add($"{fontName}:\n{val}\n");
-                    else
-                        lines.Add($"{fontName}:\nNot specified\n");
-                }
-                return string.Join("\n", lines).Trim();
-            }
-
-            string finalCopyright = MergeField(CanvasFontInformation.CopyrightNotice);
-            string finalTrademark = MergeField(CanvasFontInformation.Trademark);
-            string finalManufacturer = MergeField(CanvasFontInformation.Manufacturer);
-            string finalDesigner = MergeField(CanvasFontInformation.Designer);
-            string finalVendorUrl = MergeField(CanvasFontInformation.FontVendorUrl);
-            string finalDesignerUrl = MergeField(CanvasFontInformation.DesignerUrl);
-            string finalLicenseDesc = MergeField(CanvasFontInformation.LicenseDescription);
-            string finalLicenseUrl = MergeField(CanvasFontInformation.LicenseInfoUrl);
-
-            List<string> fontNames = metadataFonts.Select(f => $"{f.FamilyName} {f.PreferredName} ({f.TryGetInfo(CanvasFontInformation.VersionStrings)?.Value})".Trim()).ToList();
-            string firstLine = "";
-            if (metadataFonts.Count > 1)
-                firstLine = $"This font was created as a merged subset of the following fonts:\n{string.Join("\n", fontNames)}\n";
-            else if (metadataFonts.Count == 1)
-                firstLine = $"This font is a subset of {fontNames[0]}";
-            else
-                firstLine = "This font was created from SVG files.";
-
-            List<string> descList = new();
-            foreach (CMFontFace font in metadataFonts)
-            {
-                string ds = font.TryGetInfo(CanvasFontInformation.Description)?.Value;
-                if (!string.IsNullOrEmpty(ds))
-                    descList.Add($"{font.FamilyName} {font.PreferredName}: {ds}");
-            }
-            string finalDescription = firstLine;
-            if (descList.Count > 0)
-                finalDescription = firstLine + "\n\n" + string.Join("\n\n", descList);
-
-            finalDescription = finalDescription.Trim();
-
-            byte[] nameData = templateFont.Face.GetFontTable("name");
-            byte[] newNameTable = null;
-            if (nameData != null)
-            {
-                // Automatically generate a preview string
-                string finalPreview = null;
-                if (opts.generatePreviewString)
-                {
-                    List<string> previewChars = new();
-                    foreach (FontGlyph c in characters)
-                    {
-                        uint unicode = c.Character.UnicodeIndex;
-                        if (unicode < 32 || Char.IsWhiteSpace((char)unicode))
-                            continue;
-
-                        try
-                        {
-                            previewChars.Add(char.ConvertFromUtf32((int)unicode));
-                        }
-                        catch { }
-
-                        if (previewChars.Count >= 50) // TODO: make this an arg? a const?
-                            break;
-                    }
-                    finalPreview = string.Join(string.Empty, previewChars);
-                }
-
-
-                try
-                {
-                    newNameTable = RebuildNameTable(
-                        nameData,
-                        desiredFamilyName,
-                        finalCopyright,
-                        finalDescription,
-                        finalDesigner,
-                        finalManufacturer,
-                        finalTrademark,
-                        finalVendorUrl,
-                        finalDesignerUrl,
-                        finalLicenseDesc,
-                        finalLicenseUrl,
-                        opts.DesiredVersion,
-                        finalPreview);
-                }
-                catch
-                {
-                    newNameTable = nameData;
-                }
-            }
-
-            // Merge OS/2 embedding rights and ranges
-            ushort finalFsType = 0;
-            uint r1 = 0;
-            uint r2 = 0;
-            uint r3 = 0;
-            uint r4 = 0;
-            uint c1 = 0;
-            uint c2 = 0;
-
-            if (uniqueFonts.Count > 0)
-            {
-                foreach (CMFontFace font in uniqueFonts)
-                {
-                    finalFsType = GetMostRestrictiveFsType(finalFsType, fontFsTypes[font]);
-
-                    byte[] fontOs2 = font.Face.GetFontTable("OS/2");
-                    if (fontOs2 != null)
-                    {
-                        if (fontOs2.Length >= 58)
-                        {
-                            uint fr1 = (uint)((fontOs2[42] << 24) | (fontOs2[43] << 16) | (fontOs2[44] << 8) | fontOs2[45]);
-                            uint fr2 = (uint)((fontOs2[46] << 24) | (fontOs2[47] << 16) | (fontOs2[48] << 8) | fontOs2[49]);
-                            uint fr3 = (uint)((fontOs2[50] << 24) | (fontOs2[51] << 16) | (fontOs2[52] << 8) | fontOs2[53]);
-                            uint fr4 = (uint)((fontOs2[54] << 24) | (fontOs2[55] << 16) | (fontOs2[56] << 8) | fontOs2[57]);
-                            r1 |= fr1;
-                            r2 |= fr2;
-                            r3 |= fr3;
-                            r4 |= fr4;
-                        }
-                        if (fontOs2.Length >= 86)
-                        {
-                            uint fc1 = (uint)((fontOs2[78] << 24) | (fontOs2[79] << 16) | (fontOs2[80] << 8) | fontOs2[81]);
-                            uint fc2 = (uint)((fontOs2[82] << 24) | (fontOs2[83] << 16) | (fontOs2[84] << 8) | fontOs2[85]);
-                            c1 |= fc1;
-                            c2 |= fc2;
-                        }
-                    }
-                }
-            }
-            else
-            {
-                // For SVG-only subset, keep fallback font's (Segoe UI) ranges
-                byte[] fontOs2 = templateFont.Face.GetFontTable("OS/2");
-                if (fontOs2 != null)
-                {
-                    if (fontOs2.Length >= 58)
-                    {
-                        r1 = (uint)((fontOs2[42] << 24) | (fontOs2[43] << 16) | (fontOs2[44] << 8) | fontOs2[45]);
-                        r2 = (uint)((fontOs2[46] << 24) | (fontOs2[47] << 16) | (fontOs2[48] << 8) | fontOs2[49]);
-                        r3 = (uint)((fontOs2[50] << 24) | (fontOs2[51] << 16) | (fontOs2[52] << 8) | fontOs2[53]);
-                        r4 = (uint)((fontOs2[54] << 24) | (fontOs2[55] << 16) | (fontOs2[56] << 8) | fontOs2[57]);
-                    }
-                    if (fontOs2.Length >= 86)
-                    {
-                        c1 = (uint)((fontOs2[78] << 24) | (fontOs2[79] << 16) | (fontOs2[80] << 8) | fontOs2[81]);
-                        c2 = (uint)((fontOs2[82] << 24) | (fontOs2[83] << 16) | (fontOs2[84] << 8) | fontOs2[85]);
-                    }
-                }
-            }
-
-            // Fetch template tables for modification and output
-            byte[] headData = templateFont.Face.GetFontTable("head");
-            if (headData == null) throw new InvalidDataException("Missing head table in template font");
-            headData[50] = 0;
-            headData[51] = 1; // 32-bit loca format
-
-            // Try parsing version string to update head.fontRevision
-            try
-            {
-                string cleanVersion = opts.DesiredVersion ?? "Version 1.00";
-                if (cleanVersion.StartsWith("Version ", StringComparison.OrdinalIgnoreCase))
-                    cleanVersion = cleanVersion.Substring(8).Trim();
-
-                if (double.TryParse(cleanVersion, System.Globalization.NumberStyles.Any, System.Globalization.CultureInfo.InvariantCulture, out double verNum))
-                {
-                    int major = (int)verNum;
-                    int minor = (int)Math.Round((verNum - major) * 65536.0);
-                    uint fixedVer = (uint)((major << 16) | (minor & 0xFFFF));
-
-                    headData[4] = (byte)(fixedVer >> 24);
-                    headData[5] = (byte)((fixedVer >> 16) & 0xFF);
-                    headData[6] = (byte)((fixedVer >> 8) & 0xFF);
-                    headData[7] = (byte)(fixedVer & 0xFF);
-                }
-            }
-            catch { }
-
+            // 6.4. Update hhea table with new glyph count
             byte[] hheaData = templateFont.Face.GetFontTable("hhea");
             if (hheaData == null) throw new InvalidDataException("Missing hhea table in template font");
             hheaData[34] = (byte)(outputNumGlyphs >> 8);
             hheaData[35] = (byte)(outputNumGlyphs & 0xFF);
 
-            // Retrieve and update the OS/2 table of the template font.
-            // We modify the table to merge and reflect the combined properties of all source fonts:
-            // - fsType (embedding restrictions)
-            // - ulUnicodeRange (supported Unicode character ranges)
-            // - ulCodePageRange (supported code pages)
-            byte[] os2Data = templateFont.Face.GetFontTable("OS/2");
-            if (os2Data == null) throw new InvalidDataException("Missing OS/2 table in template font");
-
-            // Update fsType (licensing/embedding settings) at offset 8
-            if (os2Data.Length >= 10)
-            {
-                os2Data[8] = (byte)(finalFsType >> 8);
-                os2Data[9] = (byte)(finalFsType & 0xFF);
-            }
-
-            // Manually update ulUnicodeRange1-4 (supported Unicode ranges) at offset 42 (16 bytes)
-            if (os2Data.Length >= 58)
-            {
-                os2Data[42] = (byte)(r1 >> 24);
-                os2Data[43] = (byte)((r1 >> 16) & 0xFF);
-                os2Data[44] = (byte)((r1 >> 8) & 0xFF);
-                os2Data[45] = (byte)(r1 & 0xFF);
-
-                os2Data[46] = (byte)(r2 >> 24);
-                os2Data[47] = (byte)((r2 >> 16) & 0xFF);
-                os2Data[48] = (byte)((r2 >> 8) & 0xFF);
-                os2Data[49] = (byte)(r2 & 0xFF);
-
-                os2Data[50] = (byte)(r3 >> 24);
-                os2Data[51] = (byte)((r3 >> 16) & 0xFF);
-                os2Data[52] = (byte)((r3 >> 8) & 0xFF);
-                os2Data[53] = (byte)(r3 & 0xFF);
-
-                os2Data[54] = (byte)(r4 >> 24);
-                os2Data[55] = (byte)((r4 >> 16) & 0xFF);
-                os2Data[56] = (byte)((r4 >> 8) & 0xFF);
-                os2Data[57] = (byte)(r4 & 0xFF);
-            }
-
-            // Manually update ulCodePageRange1-2 (supported codepages) at offset 78 (8 bytes)
-            if (os2Data.Length >= 86)
-            {
-                os2Data[78] = (byte)(c1 >> 24);
-                os2Data[79] = (byte)((c1 >> 16) & 0xFF);
-                os2Data[80] = (byte)((c1 >> 8) & 0xFF);
-                os2Data[81] = (byte)(c1 & 0xFF);
-
-                os2Data[82] = (byte)(c2 >> 24);
-                os2Data[83] = (byte)((c2 >> 16) & 0xFF);
-                os2Data[84] = (byte)((c2 >> 8) & 0xFF);
-                os2Data[85] = (byte)(c2 & 0xFF);
-            }
-
-            byte[] postData = BuildPostTable(characters, outputNumGlyphs);
-
-            // Rebuild head, hhea, maxp tables
+            // 6.5. Rebuild maxp tables
             byte[] maxpData = new byte[32];
             maxpData[1] = 1; // version 1.0 (0x00010000)
             maxpData[4] = (byte)(outputNumGlyphs >> 8);
@@ -626,17 +331,7 @@ public class FontSubsetter
             maxpData[9] = (byte)(maxContoursOfAll & 0xFF);
             maxpData[15] = 1; // maxZones = 1
 
-            // Rebuild loca data
-            byte[] newLocaData;
-            using (MemoryStream locaMs = new())
-            using (BinaryWriter locaBw = new(locaMs))
-            {
-                for (int i = 0; i <= outputNumGlyphs; i++)
-                    WriteUInt32BE(locaBw, newLoca[i]);
-                newLocaData = locaMs.ToArray();
-            }
-
-            // Write all tables into the final Stream
+            // 6.6. Assign all tables to output dictionary
             Dictionary<string, byte[]> outputTables = new(StringComparer.Ordinal);
             outputTables["cmap"] = newCmapTable;
             outputTables["glyf"] = newGlyfData;
@@ -648,65 +343,16 @@ public class FontSubsetter
             if (newNameTable != null) outputTables["name"] = newNameTable;
             if (os2Data != null) outputTables["OS/2"] = os2Data;
             if (postData != null) outputTables["post"] = postData;
-            outputTables[SUBSETTER_TAG] = Encoding.ASCII.GetBytes(SUBSETTER_SIGNATURE);
+            outputTables[CMSVTable.TAG] = cmTable.Encode();
 
-            ushort numTables = (ushort)outputTables.Count;
-            int maxPower2 = 1;
-            while (maxPower2 * 2 <= numTables) maxPower2 *= 2;
-            ushort searchRange = (ushort)(maxPower2 * 16);
-            ushort entrySelector = (ushort)Math.Log(maxPower2, 2);
-            ushort rangeShift = (ushort)(numTables * 16 - searchRange);
 
+            // 7. Write the final font file to disk
             StorageFile file = opts.OutputFile
-                ?? await StorageHelper.CreateTempFileAsync($"SS\\{desiredFamilyName}.ttf").AsTask().ConfigureAwait(false);
+                ?? await StorageHelper.CreateTempFileAsync($"SS\\{opts.DesiredName}.ttf").AsTask().ConfigureAwait(false);
 
             using (Stream outStream = await file.OpenStreamForWriteAsync().ConfigureAwait(false))
-            using (BinaryWriter bw = new(outStream))
             {
-                outStream.SetLength(0);
-                WriteUInt32BE(bw, 0x00010000); // sfntVersion
-                WriteUInt16BE(bw, numTables);
-                WriteUInt16BE(bw, searchRange);
-                WriteUInt16BE(bw, entrySelector);
-                WriteUInt16BE(bw, rangeShift);
-
-                long tableRecordsPos = outStream.Position;
-                List<string> sortedTags = outputTables.Keys.OrderBy(t => t).ToList();
-                foreach (string tag in sortedTags)
-                {
-                    bw.Write(Encoding.ASCII.GetBytes(tag));
-                    WriteUInt32BE(bw, 0); // Checksum placeholder
-                    WriteUInt32BE(bw, 0); // Offset placeholder
-                    WriteUInt32BE(bw, 0); // Length placeholder
-                }
-
-                Dictionary<string, uint> offsets = new();
-                Dictionary<string, uint> lengths = new();
-                Dictionary<string, uint> checksums = new();
-
-                foreach (string tag in sortedTags)
-                {
-                    // Align to 4 bytes
-                    while (outStream.Position % 4 != 0)
-                    {
-                        bw.Write((byte)0);
-                    }
-                    offsets[tag] = (uint)outStream.Position;
-                    byte[] data = outputTables[tag];
-                    lengths[tag] = (uint)data.Length;
-                    bw.Write(data);
-                    checksums[tag] = ComputeTableChecksum(data);
-                }
-
-                // Write table directory records
-                outStream.Position = tableRecordsPos;
-                foreach (string tag in sortedTags)
-                {
-                    outStream.Position += 4; // Skip tag
-                    WriteUInt32BE(bw, checksums[tag]);
-                    WriteUInt32BE(bw, offsets[tag]);
-                    WriteUInt32BE(bw, lengths[tag]);
-                }
+                SfntWriter.WriteSfntFile(outStream, outputTables);
             }
 
             return file;
@@ -721,455 +367,7 @@ public class FontSubsetter
 
 
 
-    // ---------------------------
-    // Helpers
-    // ---------------------------
 
-    private static void Align4(BinaryWriter bw)
-    {
-        long pad = (4 - (bw.BaseStream.Position & 3)) & 3;
-        for (int i = 0; i < pad; i++)
-            bw.Write((byte)0);
-    }
-
-    private static uint ComputeTableChecksum(byte[] data)
-    {
-        uint sum = 0;
-        int i = 0;
-        while (i < data.Length)
-        {
-            uint value = 0;
-            for (int b = 0; b < 4; b++)
-            {
-                value <<= 8;
-                if (i < data.Length)
-                    value |= data[i++];
-            }
-            sum += value;
-        }
-        return sum;
-    }
-
-    // ---------------------------
-    // post table build (format 2.0)
-    // ---------------------------
-    private static byte[] BuildPostTable(IList<FontGlyph> characters, ushort outputNumGlyphs)
-    {
-        using MemoryStream ms = new();
-        using BinaryWriter bw = new(ms, Encoding.UTF8, leaveOpen: true);
-
-        // Format 2.0 (0x00020000)
-        WriteUInt32BE(bw, 0x00020000);
-        WriteUInt32BE(bw, 0); // italicAngle
-        WriteInt16BE(bw, 0);  // underlinePosition
-        WriteInt16BE(bw, 0);  // underlineThickness
-        WriteUInt32BE(bw, 0); // isFixedPitch
-        WriteUInt32BE(bw, 0); // minMemType42
-        WriteUInt32BE(bw, 0); // maxMemType42
-        WriteUInt32BE(bw, 0); // minMemType1
-        WriteUInt32BE(bw, 0); // maxMemType1
-
-        WriteUInt16BE(bw, outputNumGlyphs);
-
-        // GID 0 is .notdef
-        WriteUInt16BE(bw, 0);
-
-        // Pass 1: Write index array for all glyphs
-        ushort currentStringIndex = 258;
-        for (int i = 0; i < characters.Count; i++)
-        {
-            if (string.IsNullOrWhiteSpace(characters[i].GlyphName))
-            {
-                WriteUInt16BE(bw, 0);
-                continue;
-            }
-
-            WriteUInt16BE(bw, currentStringIndex++);
-        }
-
-        // Pass 2: Write Pascal strings directly to BinaryWriter
-        for (int i = 0; i < characters.Count; i++)
-        {
-            string name = characters[i].GlyphName;
-            if (string.IsNullOrWhiteSpace(name))
-                continue;
-
-            if (name.Length > 255)
-                name = name.Substring(0, 255);
-
-            byte[] bytes = Encoding.ASCII.GetBytes(name);
-            bw.Write((byte)bytes.Length);
-            bw.Write(bytes);
-        }
-
-        return ms.ToArray();
-    }
-
-    // ---------------------------
-    // cmap build (format 4 + 12)
-    // ---------------------------
-
-    // --------------------------------------------------------------------------------
-    // Builds the standard OpenType 'cmap' table supporting both Format 4 (Unicode BMP)
-    // and Format 12 (UCS-4 / Extended Unicode) to map characters to glyph indices.
-    // --------------------------------------------------------------------------------
-    private static byte[] BuildCmapTable(Dictionary<uint, uint> map)
-    {
-        Dictionary<uint, uint> bmpMap = map.Where(kv => kv.Key <= 0xFFFF).ToDictionary(kv => kv.Key, kv => kv.Value);
-        Dictionary<uint, uint> ucs4Map = map.Where(kv => kv.Key > 0xFFFF).ToDictionary(kv => kv.Key, kv => kv.Value);
-
-        byte[] format4 = bmpMap.Count > 0 ? BuildCmapFormat4(bmpMap) : null;
-        byte[] format12 = (bmpMap.Count > 0 || ucs4Map.Count > 0) ? BuildCmapFormat12(map) : null;
-
-        using MemoryStream ms = new();
-        using BinaryWriter bw = new(ms, Encoding.UTF8, leaveOpen: true);
-
-        ushort numSubtables = 0;
-        if (format4 != null) numSubtables++;
-        if (format12 != null) numSubtables++;
-
-        WriteUInt16BE(bw, 0);            // cmap Table Version (always 0)
-        WriteUInt16BE(bw, numSubtables); // Number of encoding subtables
-
-        long encodingRecordsPos = ms.Position;
-
-        // Write encoding record placeholders
-        if (format4 != null)
-        {
-            WriteUInt16BE(bw, 3); // platformID: Windows
-            WriteUInt16BE(bw, 1); // encodingID: Unicode BMP (UCS-2)
-            WriteUInt32BE(bw, 0); // Offset from start of cmap (placeholder)
-        }
-        if (format12 != null)
-        {
-            WriteUInt16BE(bw, 3);  // platformID: Windows
-            WriteUInt16BE(bw, 10); // encodingID: Unicode Full (UCS-4)
-            WriteUInt32BE(bw, 0);  // Offset from start of cmap (placeholder)
-        }
-
-        List<uint> offsets = [];
-
-        // Write the subtable payloads and record their actual offsets
-        if (format4 != null)
-        {
-            uint offset = (uint)ms.Position;
-            offsets.Add(offset);
-            bw.Write(format4);
-        }
-        if (format12 != null)
-        {
-            uint offset = (uint)ms.Position;
-            offsets.Add(offset);
-            bw.Write(format12);
-        }
-
-        // Backtrack and fill in the correct offsets in encoding records
-        ms.Position = encodingRecordsPos;
-        int idx = 0;
-        if (format4 != null)
-        {
-            WriteUInt16BE(bw, 3);
-            WriteUInt16BE(bw, 1);
-            WriteUInt32BE(bw, offsets[idx++]);
-        }
-        if (format12 != null)
-        {
-            WriteUInt16BE(bw, 3);
-            WriteUInt16BE(bw, 10);
-            WriteUInt32BE(bw, offsets[idx++]);
-        }
-
-        return ms.ToArray();
-    }
-
-    // --------------------------------------------------------------------------------
-    // Builds a Format 4 'cmap' subtable, which is used for character codes <= 0xFFFF.
-    // Segmented into contiguous ranges of character codes mapping to glyph indices.
-    // --------------------------------------------------------------------------------
-    private static byte[] BuildCmapFormat4(Dictionary<uint, uint> map)
-    {
-        ushort[] codes = map.Keys.Where(k => k <= 0xFFFF).Select(k => (ushort)k).OrderBy(x => x).ToArray();
-        if (codes.Length == 0)
-            codes = [(ushort)0];
-
-        // Format 4 uses segCount where each segment represents a range of characters.
-        // We use segments of size 1 (startCode == endCode) for simplicity and correctness.
-        // The last segment must be the sentinel segment mapping 0xFFFF to GID 0.
-        int segCount = codes.Length + 1;
-
-        using MemoryStream ms = new();
-        using BinaryWriter bw = new(ms, Encoding.UTF8, leaveOpen: true);
-
-        WriteUInt16BE(bw, 4); // Format number: 4
-
-        long lengthPos = ms.Position;
-        WriteUInt16BE(bw, 0); // Byte length of this subtable (placeholder)
-
-        WriteUInt16BE(bw, 0); // Language (0 for language-independent)
-
-        WriteUInt16BE(bw, (ushort)(segCount * 2)); // segCountX2
-
-        // Fast search parameters calculation
-        int pow2 = 1;
-        while (pow2 * 2 <= segCount) pow2 *= 2;
-        ushort searchRange = (ushort)(pow2 * 2);
-        ushort entrySelector = (ushort)Math.Log(pow2, 2);
-        ushort rangeShift = (ushort)(segCount * 2 - searchRange);
-
-        WriteUInt16BE(bw, searchRange);
-        WriteUInt16BE(bw, entrySelector);
-        WriteUInt16BE(bw, rangeShift);
-
-        // endCode[segCount]
-        foreach (ushort c in codes)
-            WriteUInt16BE(bw, c);
-        WriteUInt16BE(bw, 0xFFFF); // Sentinel endCode
-
-        // reservedPad (always 0)
-        WriteUInt16BE(bw, 0);
-
-        // startCode[segCount]
-        foreach (ushort c in codes)
-            WriteUInt16BE(bw, c);
-        WriteUInt16BE(bw, 0xFFFF); // Sentinel startCode
-
-        // idDelta[segCount]
-        // idDelta[i] = (glyphId - characterCode) % 65536
-        foreach (ushort c in codes)
-        {
-            uint gid = map[c];
-            ushort diff = (ushort)(gid - c);
-            short delta = (short)diff; // Modulo arithmetic
-            WriteInt16BE(bw, delta);
-        }
-        WriteInt16BE(bw, 1); // Sentinel segment delta
-
-        // idRangeOffset[segCount] (0 because we map directly using idDelta)
-        for (int i = 0; i < segCount; i++)
-            WriteUInt16BE(bw, 0);
-
-        long endPos = ms.Position;
-        ms.Position = lengthPos;
-        WriteUInt16BE(bw, (ushort)endPos);
-        ms.Position = endPos;
-
-        return ms.ToArray();
-    }
-
-    // --------------------------------------------------------------------------------
-    // Builds a Format 12 'cmap' subtable, supporting 32-bit UCS-4 character codes.
-    // --------------------------------------------------------------------------------
-    private static byte[] BuildCmapFormat12(Dictionary<uint, uint> map)
-    {
-        uint[] codes = map.Keys.OrderBy(k => k).ToArray();
-        if (codes.Length == 0)
-            codes = [0u];
-
-        using MemoryStream ms = new();
-        using BinaryWriter bw = new(ms, Encoding.UTF8, leaveOpen: true);
-
-        WriteUInt16BE(bw, 12); // Format number: 12
-        WriteUInt16BE(bw, 0);  // Reserved (must be 0)
-
-        long lengthPos = ms.Position;
-        WriteUInt32BE(bw, 0);  // Byte length of subtable (placeholder)
-
-        WriteUInt32BE(bw, 0);  // Language (0 for language-independent)
-
-        uint nGroups = (uint)codes.Length;
-        WriteUInt32BE(bw, nGroups);
-
-        // Groups representation (startCharCode, endCharCode, startGlyphId)
-        foreach (uint c in codes)
-        {
-            uint gid = map[c];
-            WriteUInt32BE(bw, c);   // startCharCode
-            WriteUInt32BE(bw, c);   // endCharCode
-            WriteUInt32BE(bw, gid); // startGlyphId
-        }
-
-        long endPos = ms.Position;
-        uint length = (uint)endPos;
-        ms.Position = lengthPos;
-        WriteUInt32BE(bw, length);
-        ms.Position = endPos;
-
-        return ms.ToArray();
-    }
-
-    private static byte[] RebuildNameTable(
-        byte[] originalNameData, 
-        string mergedName, 
-        string finalCopyright, 
-        string finalDescription, 
-        string finalDesigner, 
-        string finalManufacturer,
-        string finalTrademark,
-        string finalVendorUrl,
-        string finalDesignerUrl,
-        string finalLicenseDesc,
-        string finalLicenseUrl,
-        string finalVersion,
-        string finalPreview)
-    {
-        using MemoryStream msIn = new(originalNameData);
-        using BinaryReader br = new(msIn);
-        
-        ushort format = ReadUInt16BE(br);
-        ushort count = ReadUInt16BE(br);
-        ushort stringOffset = ReadUInt16BE(br);
-        
-        List<(ushort PlatformID, ushort EncodingID, ushort LanguageID, ushort NameID, string Value)> records = new();
-        
-        for (int i = 0; i < count; i++)
-        {
-            ushort platformID = ReadUInt16BE(br);
-            ushort encodingID = ReadUInt16BE(br);
-            ushort languageID = ReadUInt16BE(br);
-            ushort nameID = ReadUInt16BE(br);
-            ushort length = ReadUInt16BE(br);
-            ushort offset = ReadUInt16BE(br);
-            
-            long savePos = msIn.Position;
-            msIn.Position = stringOffset + offset;
-            byte[] stringBytes = br.ReadBytes(length);
-            msIn.Position = savePos;
-            
-            string val = "";
-            if (platformID == 3 || platformID == 0)
-            {
-                byte[] temp = new byte[stringBytes.Length];
-                for (int j = 0; j < temp.Length; j += 2)
-                {
-                    if (j + 1 < temp.Length)
-                    {
-                        temp[j] = stringBytes[j+1];
-                        temp[j+1] = stringBytes[j];
-                    }
-                }
-                val = Encoding.Unicode.GetString(temp);
-            }
-            else
-            {
-                val = Encoding.ASCII.GetString(stringBytes);
-            }
-            
-            records.Add((platformID, encodingID, languageID, nameID, val));
-        }
-
-        bool hasCopyright = records.Any(r => r.NameID == 0);
-        bool hasDescription = records.Any(r => r.NameID == 10);
-        bool hasDesigner = records.Any(r => r.NameID == 9);
-        bool hasManufacturer = records.Any(r => r.NameID == 8);
-        bool hasTrademark = records.Any(r => r.NameID == 7);
-        bool hasVendorUrl = records.Any(r => r.NameID == 11);
-        bool hasDesignerUrl = records.Any(r => r.NameID == 12);
-        bool hasLicenseDesc = records.Any(r => r.NameID == 13);
-        bool hasLicenseUrl = records.Any(r => r.NameID == 14);
-        bool hasVersion = records.Any(r => r.NameID == 5);
-        bool hasPreview = records.Any(r => r.NameID == 19);
-        
-        if (!hasCopyright && !string.IsNullOrEmpty(finalCopyright))
-            records.Add((3, 1, 1033, 0, finalCopyright));
-        if (!hasDescription && !string.IsNullOrEmpty(finalDescription))
-            records.Add((3, 1, 1033, 10, finalDescription));
-        if (!hasDesigner && !string.IsNullOrEmpty(finalDesigner))
-            records.Add((3, 1, 1033, 9, finalDesigner));
-        if (!hasManufacturer && !string.IsNullOrEmpty(finalManufacturer))
-            records.Add((3, 1, 1033, 8, finalManufacturer));
-        if (!hasTrademark && !string.IsNullOrEmpty(finalTrademark))
-            records.Add((3, 1, 1033, 7, finalTrademark));
-        if (!hasVendorUrl && !string.IsNullOrEmpty(finalVendorUrl))
-            records.Add((3, 1, 1033, 11, finalVendorUrl));
-        if (!hasDesignerUrl && !string.IsNullOrEmpty(finalDesignerUrl))
-            records.Add((3, 1, 1033, 12, finalDesignerUrl));
-        if (!hasLicenseDesc && !string.IsNullOrEmpty(finalLicenseDesc))
-            records.Add((3, 1, 1033, 13, finalLicenseDesc));
-        if (!hasLicenseUrl && !string.IsNullOrEmpty(finalLicenseUrl))
-            records.Add((3, 1, 1033, 14, finalLicenseUrl));
-        if (!hasVersion && !string.IsNullOrEmpty(finalVersion))
-            records.Add((3, 1, 1033, 5, finalVersion));
-        if (!hasPreview && !string.IsNullOrEmpty(finalPreview))
-            records.Add((3, 1, 1033, 19, finalPreview));
-        
-        using MemoryStream msOut = new();
-        using BinaryWriter bw = new(msOut);
-        
-        WriteUInt16BE(bw, format);
-        WriteUInt16BE(bw, (ushort)records.Count);
-        
-        long stringOffsetPos = msOut.Position;
-        WriteUInt16BE(bw, 0);
-        
-        using MemoryStream stringMs = new();
-        using BinaryWriter stringBw = new(stringMs);
-        
-        foreach ((ushort PlatformID, ushort EncodingID, ushort LanguageID, ushort NameID, string Value) rec in records)
-        {
-            string newValue = rec.Value;
-            if (rec.NameID == 1 || rec.NameID == 4 || rec.NameID == 16)
-                newValue = mergedName;
-            else if (rec.NameID == 6)
-                newValue = mergedName.Replace(" ", "");
-            else if (rec.NameID == 0)
-                newValue = finalCopyright;
-            else if (rec.NameID == 5)
-                newValue = finalVersion;
-            else if (rec.NameID == 7)
-                newValue = finalTrademark;
-            else if (rec.NameID == 8)
-                newValue = finalManufacturer;
-            else if (rec.NameID == 9)
-                newValue = finalDesigner;
-            else if (rec.NameID == 10)
-                newValue = finalDescription;
-            else if (rec.NameID == 11)
-                newValue = finalVendorUrl;
-            else if (rec.NameID == 12)
-                newValue = finalDesignerUrl;
-            else if (rec.NameID == 13)
-                newValue = finalLicenseDesc;
-            else if (rec.NameID == 14)
-                newValue = finalLicenseUrl;
-            else if (rec.NameID == 19)
-                newValue = finalPreview;
-            
-            byte[] bytes;
-            if (rec.PlatformID == 3 || rec.PlatformID == 0)
-            {
-                byte[] utf16 = Encoding.Unicode.GetBytes(newValue);
-                bytes = new byte[utf16.Length];
-                for (int j = 0; j < utf16.Length; j += 2)
-                {
-                    bytes[j] = utf16[j+1];
-                    bytes[j+1] = utf16[j];
-                }
-            }
-            else
-            {
-                bytes = Encoding.ASCII.GetBytes(newValue);
-            }
-            
-            ushort length = (ushort)bytes.Length;
-            ushort offset = (ushort)stringMs.Position;
-            stringBw.Write(bytes);
-            
-            WriteUInt16BE(bw, rec.PlatformID);
-            WriteUInt16BE(bw, rec.EncodingID);
-            WriteUInt16BE(bw, rec.LanguageID);
-            WriteUInt16BE(bw, rec.NameID);
-            WriteUInt16BE(bw, length);
-            WriteUInt16BE(bw, offset);
-        }
-        
-        ushort finalStringOffset = (ushort)msOut.Position;
-        bw.Write(stringMs.ToArray());
-        
-        msOut.Position = stringOffsetPos;
-        WriteUInt16BE(bw, finalStringOffset);
-        
-        return msOut.ToArray();
-    }
 
     // --------------------------------------------------------------------------------
     // Extracts a glyph's outline geometry (supporting both TTF and CFF/PostScript) 
@@ -1501,16 +699,9 @@ public class FontSubsetter
         return ms.ToArray();
     }
 
-    private static ushort GetMostRestrictiveFsType(ushort fsType1, ushort fsType2)
-    {
-        if ((fsType1 & 0x0002) != 0 || (fsType2 & 0x0002) != 0)
-            return 0x0002;
-        if ((fsType1 & 0x0004) != 0 || (fsType2 & 0x0004) != 0)
-            return 0x0004;
-        if ((fsType1 & 0x0008) != 0 || (fsType2 & 0x0008) != 0)
-            return 0x0008;
-        return 0;
-    }
+ 
+
+
 }
 
 public class TrueTypeGlyphReceiver : ICanvasPathReceiver
